Unit tests

+++ BoardGameEventBackend/src/main/java/com/example/BoardGameEventBackend/controller/EventController.java
@@ -2,10 +2,19 @@ package com.example.BoardGameEventBackend.controller;
 
 
 import com.example.BoardGameEventBackend.dto.*;
+import com.example.BoardGameEventBackend.exception.EventNotFoundException;
+import com.example.BoardGameEventBackend.exception.ForbiddenException;
 import com.example.BoardGameEventBackend.model.Event;
+import com.example.BoardGameEventBackend.model.User;
 import com.example.BoardGameEventBackend.service.EventService;
+import com.example.BoardGameEventBackend.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -17,6 +26,7 @@ import java.util.List;
 @RequestMapping("/api")
 public class EventController {
     private final EventService eventService;
+    private final UserService userService;
 
     @GetMapping("/events")
     public List<EventDto> getAllGameEvents(){
@@ -37,6 +47,16 @@ public class EventController {
     @PutMapping("/events/{id}")
     @PreAuthorize("hasAuthority('ADMIN')")
     public EventDto updateGameEvent(@PathVariable Long id, @Valid @RequestBody Event event){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails principal = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findByUsername(principal.getUsername());
+
+        if(authentication.getAuthorities().stream().map(GrantedAuthority::getAuthority).noneMatch(role -> role.contains("ADMIN"))
+                && !loggedUser.getId().equals(id)){
+            throw new ForbiddenException();
+        }
+
         return EventDtoMapper.mapToEventDto(eventService.updateGameEvent(id, event));
     }
 
@@ -55,6 +75,18 @@ public class EventController {
     @DeleteMapping("/events/{eventId}/players/{userId}")
     @PreAuthorize("hasAuthority('USER')")
     public EventDto removePlayerToEvent(@PathVariable("eventId") Long eventId, @PathVariable("userId") Long userId){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails principal = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findByUsername(principal.getUsername());
+
+        Event event = eventService.getGameEvent(eventId);
+
+        if(authentication.getAuthorities().stream().map(GrantedAuthority::getAuthority).noneMatch(role -> role.contains("ADMIN"))
+                && !loggedUser.getId().equals(userId) && !loggedUser.getId().equals(event.getOrganizer().getId())){
+            throw new ForbiddenException();
+        }
+
         return EventDtoMapper.mapToEventDto(eventService.removePlayerToEvent(eventId, userId));
     }
